+++ CHANGELOG.md
@@ -19,10 +19,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for OpenGL ES 2.0
 - Escape sequence to set underline color (`CSI 58 : 2 : Ps : Ps : Ps m`/`CSI 58 : 5 : Ps m`)
 - Escape sequence to reset underline color (`CSI 59 m`)
+- Support for inline input method
 
 ### Changed
 
 - The `--help` output was reworked with a new colorful syntax
+- Search bar is now respecting cursor thickness
 
 ### Fixed